pyproject.toml
[project]
name = "veyra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "logfire[fastapi]>=4.3.5",
    "openai>=1.101.0",
    "pydantic-ai-slim[ag-ui,anthropic,google]>=0.7.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
